Refactor ProductInsert to build inputs with separate component. Reduce bloating in render method.
